requirements/base.txt
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5

# Django
# ------------------------------------------------------------------------------
django==5.0.7  # pyup: < 5.0  # https://www.djangoproject.com/
django-allauth[mfa]==0.63.3  # https://github.com/pennersr/django-allauth
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat
django-cors-headers==4.4.0  # https://github.com/adamchainz/django-cors-headers
django-crispy-forms==2.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-filter==23.5
django-model-utils==4.5.1  # https://github.com/jazzband/django-model-utils
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
drf-yasg==1.21.7
flower==2.0.1  # https://github.com/mher/flower
hiredis==2.3.2  # https://github.com/redis/hiredis-py
ipython==8.26.0
kafka-python-ng==2.2.3
pika==1.3.2
Pillow==10.4.0  # https://github.com/python-pillow/Pillow
python-magic==0.4.27
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
redis==5.0.7  # https://github.com/redis/redis-py
whitenoise==6.7.0  # https://github.com/evansd/whitenoise
